Added Authotization and fix some issue and added auto login facilities

+++ backend/routes/posts.js
@@ -61,7 +61,8 @@ router.post('', checkAuth, multer({storage: storage}).single("image"), (req, res
     const post = new Post({
         title: req.body.title,
         content: req.body.content,
-        imagePath: url + "/images/" + req.file.filename
+        imagePath: url + "/images/" + req.file.filename,
+        creator: req.userData.userId,
     })
     post.save().then((createdPost) => {
         res.status(201).json({
@@ -71,6 +72,7 @@ router.post('', checkAuth, multer({storage: storage}).single("image"), (req, res
                 title: createdPost.title,
                 content: createdPost.content,
                 imagePath: createdPost.imagePath,
+                creator: createdPost.creator
             }
         })
     }).catch((error) => {
@@ -116,16 +118,22 @@ router.put('/:id', checkAuth, multer({storage: storage}).single("image"), (req, 
         content: req.body.content,
         imagePath: imagePath,
     })
-    Post.updateOne({_id: req.params.id}, post).then((updatedPost) => {
-        res.status(200).json({
-            message: "Posts updated successfully!",
-            post: {
-                id: updatedPost._id,
-                title: updatedPost.title,
-                content: updatedPost.content,
-                imagePath: updatedPost.imagePath,
-            }
-        })
+    Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then((updatedPost) => {
+        if (updatedPost.modifiedCount > 0 || updatedPost.matchedCount > 0) {
+            res.status(200).json({
+                message: "Posts updated successfully!",
+                post: {
+                    id: updatedPost._id,
+                    title: updatedPost.title,
+                    content: updatedPost.content,
+                    imagePath: updatedPost.imagePath,
+                }
+            })
+        } else {
+            res.status(401).json({
+                message: "You are not authorized to perform this action!"
+            })
+        }
     }).catch((error) => {
         res.status(500).json({
             message: "Something went wrong!",
@@ -134,9 +142,15 @@ router.put('/:id', checkAuth, multer({storage: storage}).single("image"), (req, 
     })
 })
 
-router.delete('/:id', (req, res, next) => {
-    Post.deleteOne({_id: req.params.id}).then((result) => {
-        res.status(200).json({ message: 'Post deleted!' })
+router.delete('/:id', checkAuth, (req, res, next) => {
+    Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then((result) => {
+        if (result.deletedCount > 0) {
+            res.status(200).json({ message: 'Post deleted!' })
+        } else {
+            res.status(401).json({
+                message: "You are not authorized to perform this action!"
+            })
+        }
     }).catch((error) => {
         res.status(500).json({
             message: "Something went wrong!",
